Fix default vlan state overridden issue (#426)
Fix default vlan state overridden issue

SUMMARY
Fixes #425
ISSUE TYPE

Bugfix Pull Request

COMPONENT NAME
nxos_vlans
ADDITIONAL INFORMATION
This fix changes the fact gathering logic for handling the unstructured output that comes back for vlan to vni mapping.  The old logic did a check to see if the vlan number exists in the output which could easily match in cases where a match is not desired.  This was causing the default vlan to be changed if 1  matched anything next to a \n character  in the output line

Reviewed-by: Nilashish Chakraborty <nilashishchakraborty8@gmail.com>
Reviewed-by: None <None>