Add `vrf` option to `nxos_nxapi` module.
Comments on the design:
This is an old module with only state: present / absent.
As far as I can see, when state: present is specified,
it does not change parameters that are not mentioned.
For example, if the nxapi sandbox command is already present on the device
but the sandbox parameter is not specified in the playbook,
it appears that no deletion or other actions are taken.

Therefore, the use-vrf follows this design.

Not applying the `nxapi use-vrf` command at all means enabling nxapi for all VRFs.
However, in my implementation of this module, there is currently no way
to remove an already applied nxapi use-vrf <xxxxx> configuration at all.
This is because while it is possible to specify another VRF,
but there is no way to reflect the absence of a parameter.
This aligns with the aforementioned behavior of state: `present`
where nothing is done if a parameter is not specified.

This implementation method was chosen to maintain consistency within the module, provide clarity,
avoid disrupting environments using the existing nxos_nxapi module, and keep the parameters simple.

Other possible approaches:
- Implementing a behavior similar to the state: overriden in other modules,
  meaning if no VRF is specified, the existing nxapi use-vrf setting would be deleted.
- Making the parameters more complex. For example, explicitly enabling or disabling as follows:
```
vrf:
  limit_to_vrf: bool
  vrf_name: xxx
```

Reference for the supported versions of use-vrf:
- For N9K:
  At least, supported in version 7.0(3)I3(1)
  https://www.cisco.com/c/en/us/td/docs/switches/datacenter/nexus9000/sw/7-x/command_references/configuration_commands/b_Using_N9K_Config_Commands/b_N9K_Bookmap_chapter_01111.html#wp4568333650
  Also, it is not supported in version 6.1(2)I2(2)
  https://www.cisco.com/c/en/us/td/docs/switches/datacenter/nexus9000/sw/6-x/command_reference/config_612I22/b_n9k_command_ref/b_n9k_command_ref_chapter_010000.html#wp2897910537
  I am unsure of the exact version where it was released. I have reviewed all the release notes, but there is no mention of it.
- For N7K:
  > The nxapi use-vrf feature is introduced in Cisco NX-OS Release 8.2(3), which helps to secure NX-API by binding the NX-API to a particular VRF.
  https://www.cisco.com/c/en/us/td/docs/switches/datacenter/nexus7000/sw/programmability/guide/cisco_nexus7000_programmability_guide_8x/b-cisco-nexus7000-programmability-guide-8x_chapter_011.html
- For N3K:
  At least, supported in version 7.0(3)I4(1)
  https://www.cisco.com/c/en/us/td/docs/switches/datacenter/nexus3000/sw/command/reference/703i74m3k/config/b_N3K_Config_Commands_703i74/b_N3K_Config_Commands_703i74_chapter_01111.html#wp1351653258