Issue #819: Fix replace LAG which has no existing members
Setting `state: replaced` on LAG interfaces when the existing config had no members resulted in an
error:

*"MODULE FAILURE\nSee stdout/stderr for the exact error"*

This happened in the `intersect_list_of_dicts` which would receive `h=None` in this scenario.

This change adds checks to the `intersect_list_of_dicts` function in the same way per `diff_list_of_dicts`.