Explicitly setting *secondary: false* on l3_interface results in change
When diffing the want and have for an IPv4 address, if the user has explicitly set secondary to
false the diff result will always be truthy and result in a "change".  Like *tag* this change sets a
the default value for secondary on both want and have such that the diff will work as expected.