Issue #988
This change does the following:

* detects where the mode is access by default and reports it in the facts
* it limits operations to interfaces which are configured as layer2 interfaces

By augmenting the facts with `mode: access` where it was not previously, the use of `mode: access`
becomes idempotent.

BREAKING:  If an input configuration is supplied that attempts to apply L2 configuration to an
interface that is L3 (or doesn't exist), an exception will be raised telling the user that they
should use `nxos_interfaces` to first configure the interface for L2.

Fixes: #988